Fixed a bug in the compiler
The compiler used to provide horribly wrong suggestions of what kind of
token could be following another token. This was caused by an old
version of TParser.token, the file used to get the token name based on
the token number, that somehow managed to survive in the generated jars.

git-svn-id: https://vcs.informatik.uni-siegen.de/svn/algvistool/trunk@956 57b514eb-d8df-4872-b74f-14c3cc28987e